@@ -16,9 +16,9 @@
 
 package com.adaptris.core.oauth.azure;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -27,8 +27,8 @@ import static org.mockito.Mockito.when;
 import java.io.IOException;
 import java.util.concurrent.Future;
 
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import com.adaptris.core.AdaptrisMessage;
 import com.adaptris.core.AdaptrisMessageFactory;
@@ -42,7 +42,7 @@ import com.microsoft.aad.adal4j.ClientCredential;
 
 public class AzureClientSecretTokenTest {
 
-  @Before
+  @BeforeEach
   public void setUp() throws Exception {
 
   }
