@@ -16,9 +16,9 @@
 
 package com.adaptris.core.oauth.generic;
 import static com.adaptris.core.oauth.generic.JsonResponseHandlerTest.ACCESS_TOKEN;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
+import org.junit.jupiter.api.Test;
 import com.adaptris.core.AdaptrisMessage;
 import com.adaptris.core.AdaptrisMessageFactory;
 import com.adaptris.core.ServiceException;
